+++ packages/offix-client/src/offline/OfflineLink.ts
@@ -2,10 +2,10 @@ import { ApolloLink, NextLink, Operation, Observable } from "apollo-link";
 import { NetworkInfo, NetworkStatus, OfflineQueueListener, OfflineMutationsHandler, OfflineStore } from ".";
 import { OfflineQueue } from "./OfflineQueue";
 import { ObjectState } from "../conflicts";
-import { OperationQueueEntry } from "./OperationQueueEntry";
 import * as debug from "debug";
 import { QUEUE_LOGGER } from "../config/Constants";
 import { OfflineError } from "./OfflineError";
+import { IResultProcessor } from "./procesors/IResultProcessor";
 
 export const logger = debug.default(QUEUE_LOGGER);
 
@@ -13,7 +13,7 @@ export interface OfflineLinkOptions {
   networkStatus: NetworkStatus;
   store: OfflineStore;
   listener?: OfflineQueueListener;
-  conflictStateProvider?: ObjectState;
+  resultProcessors?: IResultProcessor[];
 }
 
 /**
